requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
ffmpeg-python==0.1.17
future==0.17.1
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pydub==0.23.1
Pygments==2.4.0
pytesseract==0.2.6
requests==2.22.0
six==1.12.0
traitlets==4.3.2
urllib3==1.25.2
wcwidth==0.1.7
Werkzeug==0.15.4
yapf==0.27.0
